arreglado todo lo del perfil y arreglos en la pantalla de pasarela de pagos para que se miren los cambios reflejados
